+++ app/src/proto/clusters_pb_service.js
@@ -64,6 +64,15 @@ Clusters.GetApplication = {
   responseType: clusters_pb.GetApplicationResponse
 };
 
+Clusters.GetApplicationsTopology = {
+  methodName: "GetApplicationsTopology",
+  service: Clusters,
+  requestStream: false,
+  responseStream: false,
+  requestType: clusters_pb.GetApplicationsTopologyRequest,
+  responseType: clusters_pb.GetApplicationsTopologyResponse
+};
+
 exports.Clusters = Clusters;
 
 function ClustersClient(serviceHost, options) {
@@ -231,6 +240,37 @@ ClustersClient.prototype.getApplication = function getApplication(requestMessage
     callback = arguments[1];
   }
   var client = grpc.unary(Clusters.GetApplication, {
+    request: requestMessage,
+    host: this.serviceHost,
+    metadata: metadata,
+    transport: this.options.transport,
+    debug: this.options.debug,
+    onEnd: function (response) {
+      if (callback) {
+        if (response.status !== grpc.Code.OK) {
+          var err = new Error(response.statusMessage);
+          err.code = response.status;
+          err.metadata = response.trailers;
+          callback(err, null);
+        } else {
+          callback(null, response.message);
+        }
+      }
+    }
+  });
+  return {
+    cancel: function () {
+      callback = null;
+      client.close();
+    }
+  };
+};
+
+ClustersClient.prototype.getApplicationsTopology = function getApplicationsTopology(requestMessage, metadata, callback) {
+  if (arguments.length === 2) {
+    callback = arguments[1];
+  }
+  var client = grpc.unary(Clusters.GetApplicationsTopology, {
     request: requestMessage,
     host: this.serviceHost,
     metadata: metadata,